third_party/blink/web_tests/external/wpt/css/css-box/margin-trim/computed-margin-values/block-container-block-end-nested-child-expected.txt

This is a testharness.js-based test.
PASS item 1
FAIL item 2 assert_equals: 
<item data-expected-margin-bottom="0" style="block-size: auto;">
                <div><item data-expected-margin-bottom="10"></item></div>
                <div>
                    <item data-expected-margin-bottom="0" style="block-size: auto;">
                        <div><item data-expected-margin-bottom="0"></item></div>
                    </item>
                    <item class="collapsed" data-expected-margin-bottom="0">
                        <div><item class="collapsed" data-expected-margin-bottom="0"></item></div>
                    </item>
                </div>
                <item class="collapsed" data-expected-margin-bottom="0">
                    <div><item class="collapsed" data-expected-margin-bottom="0"></item></div>
                </item>
            </item>
margin-bottom expected "0" but got "10"
PASS item 3
FAIL item 4 assert_equals: 
<item data-expected-margin-bottom="0" style="block-size: auto;">
                        <div><item data-expected-margin-bottom="0"></item></div>
                    </item>
margin-bottom expected "0" but got "10"
FAIL item 5 assert_equals: 
<item data-expected-margin-bottom="0"></item>
margin-bottom expected "0" but got "10"
FAIL item 6 assert_equals: 
<item class="collapsed" data-expected-margin-bottom="0">
                        <div><item class="collapsed" data-expected-margin-bottom="0"></item></div>
                    </item>
margin-bottom expected "0" but got "10"
FAIL item 7 assert_equals: 
<item class="collapsed" data-expected-margin-bottom="0"></item>
margin-bottom expected "0" but got "10"
FAIL item 8 assert_equals: 
<item class="collapsed" data-expected-margin-bottom="0">
                    <div><item class="collapsed" data-expected-margin-bottom="0"></item></div>
                </item>
margin-bottom expected "0" but got "10"
FAIL item 9 assert_equals: 
<item class="collapsed" data-expected-margin-bottom="0"></item>
margin-bottom expected "0" but got "10"
FAIL item 10 assert_equals: 
<item class="collapsed" data-expected-margin-bottom="0">
                <div><item class="collapsed" data-expected-margin-bottom="0"></item></div>
            </item>
margin-bottom expected "0" but got "10"
FAIL item 11 assert_equals: 
<item class="collapsed" data-expected-margin-bottom="0"></item>
margin-bottom expected "0" but got "10"
Harness: the test ran to completion.
